requirements.txt
amqp==2.6.1
appdirs==1.4.4
argh==0.26.2
asgiref==3.2.3
attrs==23.1.0
beautifulsoup4==4.12.3
billiard==3.6.4.0
black==19.10b0
boto3==1.16.20
botocore==1.19.63
brotlipy==0.7.0
bs4==0.0.2
bytecode==0.14.2
cattrs==23.1.2
celery==4.4.3
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
click==8.1.3
DateTime==5.5
ddsketch==2.0.4
ddtrace==1.5.2
Django==3.0.3
django-auth-ldap==2.1.0
django-celery-beat==2.0.0
django-modelclone==0.7.1
django-money==1.0
django-mysql==3.3.0
django-redis==4.12.1
django-redis-sessions==0.6.1
django-reversion==3.0.7
django-timezone-field==4.2.3
envier==0.4.0
exceptiongroup==1.1.1
gunicorn==20.0.4
hvac==0.9.6
idna==2.9
jmespath==0.10.0
jsonschema==4.17.3
kombu==4.6.11
mpremote==1.24.1
mysql==0.0.3
mysql-connector-python==8.3.0
mysqlclient==2.1.1
packaging==23.1
pathspec==0.11.1
pathtools==0.1.2
pillow==11.1.0
protobuf==4.23.2
py-moneyed==0.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyrsistent==0.19.3
pyserial==3.5
python-crontab==2.7.1
python-dateutil==2.8.1
python-dotenv==0.12.0
python-ldap==3.2.0
pytz==2019.3
PyYAML==5.3
redis==3.4.1
regex==2023.6.3
requests==2.23.0
s3transfer==0.3.7
sentry-sdk==0.16.1
six==1.14.0
soupsieve==2.6
sqlparse==0.3.0
tenacity==8.2.2
toml==0.10.2
typed-ast==1.5.4
typing_extensions==4.6.3
urllib3==1.25.8
vine==1.3.0
watchdog==0.10.2
whitenoise==5.0.1
xmltodict==0.13.0
zope.interface==6.3
